# PERF: DataFrame._reduce