```
from asv_bench.benchmarks.groupby import *

self = GroupByMethods()
self.setup("int", "skew", "direct", 2)

%timeit self.time_dtype_as_group("int", "skew", "direct", 2)
434 ms ± 17.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
367 ms ± 8.51 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
```